Further fixed a rounding bug where the content boxes would move when you resize them. Did this by changing a bunch of width's, heights, X values, Y values, and offsets to be doubles instead of ints. Also changed the snapping tolerance to 7 instead of 5 because its fairly tricky to get a guide to snap onto an item.
My original commit before this overlooked loading a guide and it was doing a Integer.parseInt on a double value. I have fixed this.

+++ src/main/java/ca/sqlpower/wabit/dao/WorkspaceSAXHandler.java
@@ -983,7 +983,7 @@ public class WorkspaceSAXHandler extends DefaultHandler {
         	String offsetAmount = attributes.getValue("offset");
         	checkMandatory("axis", axisName);
         	checkMandatory("offset", offsetAmount);
-        	Guide guide = new Guide(Axis.valueOf(axisName), Integer.parseInt(offsetAmount));
+        	Guide guide = new Guide(Axis.valueOf(axisName), Double.parseDouble(offsetAmount));
         	if(guideName != null) {
         		guide.setName(guideName);
         	}
